Update NoResults search message
Move the useQueryParams hook outside of search
