+++ indico/modules/search/client/js/components/SearchApp.jsx
@@ -8,12 +8,11 @@
 import searchURL from 'indico-url:search.api_search';
 
 import PropTypes from 'prop-types';
-import React, {useCallback, useEffect, useMemo, useState} from 'react';
-import {useHistory} from 'react-router-dom';
-import {Grid, Loader, Menu} from 'semantic-ui-react';
+import React, {useEffect, useMemo, useState} from 'react';
+import {Grid, Loader, Menu, Message} from 'semantic-ui-react';
 
-import {useIndicoAxios} from 'indico/react/hooks';
-import {Translate} from 'indico/react/i18n';
+import {useIndicoAxios, useQueryParams} from 'indico/react/hooks';
+import {Param, Translate} from 'indico/react/i18n';
 
 import ResultList from './ResultList';
 import Category from './results/Category';
@@ -21,7 +20,6 @@ import Contribution from './results/Contribution';
 import Event from './results/Event';
 import EventNote from './results/EventNote';
 import File from './results/File';
-import NoResults from './results/NoResults';
 import SearchBar from './SearchBar';
 import SideBar from './SideBar';
 
@@ -90,35 +88,26 @@ SearchTypeMenuItem.defaultProps = {
   onClick: undefined,
 };
 
-function useQueryParams() {
-  const [query, _setQuery] = useState(window.location.search);
-  const queryObject = useMemo(() => Object.fromEntries(new URLSearchParams(query)), [query]);
-  const history = useHistory();
-
-  const setQuery = useCallback(
-    (type, name, reset) => {
-      const params = new URLSearchParams(reset ? undefined : query);
-      if (name !== undefined) {
-        params.set(type, name);
-      } else {
-        params.delete(type);
-      }
-      const _query = params.toString();
-      history.push({
-        pathname: window.location.pathname,
-        search: `?${_query}`,
-      });
-      _setQuery(_query);
-    },
-    [history, query]
+function NoResults({query}) {
+  return (
+    <Message warning={!!query}>
+      {query ? (
+        <>
+          <Message.Header>{Translate.string('No Results')}</Message.Header>
+          <Translate>
+            Your search - <Param name="query" value={query} /> - did not match any results
+          </Translate>
+        </>
+      ) : (
+        <Translate>Please enter a term above to begin searching</Translate>
+      )}
+    </Message>
   );
-
-  useEffect(() => {
-    return history.listen(location => _setQuery(location.search));
-  }, [history, query, _setQuery]);
-
-  return [queryObject, setQuery];
 }
+
+NoResults.propTypes = {
+  query: PropTypes.string.isRequired,
+};
 
 export default function SearchApp() {
   const [query, setQuery] = useQueryParams();
